requirements.txt
azure-cli-core==2.48.1
